Create qc_handler_factory
